+++ src/utils/helpers/hooks.ts
@@ -4,7 +4,7 @@ import {
   useParsedQuery,
 } from '@shlinkio/shlink-frontend-kit';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router';
 import { useSwipeable as useReactSwipeable } from 'react-swipeable';
 import type { MediaMatcher } from '../types';